bulletin_uploader/azure function version/bultenfunc/requirements.txt
azure-functions
azure-storage-blob
requests

